refactor(student-dashboard): split to smaller components
