+++ client/src/modules/StudentDashboard/components/MentorInfo/index.ts
@@ -1,0 +1,1 @@
+export { default as MentorInfo } from './MentorInfo';
