+++ client/src/modules/StudentDashboard/components/SubmitTaskSolution/SubmitTaskSolution.tsx
@@ -1,0 +1,43 @@
+import { Button, Form, Input } from 'antd';
+import React from 'react';
+import { CourseTaskSelect } from 'components/Forms';
+import { ModalSubmitForm } from 'components/Forms/ModalSubmitForm';
+import { urlPattern } from 'services/validators';
+import { useSubmitTaskSolution } from '../../hooks/useSubmitTaskSolution';
+
+interface TaskSolutionModalProps {
+  courseId: number;
+}
+
+function SubmitTaskSolution({ courseId }: TaskSolutionModalProps) {
+  const { state, closeModal, showModal, saveSolution } = useSubmitTaskSolution(courseId);
+
+  return (
+    <>
+      <Button onClick={showModal} type="primary">
+        Submit Task
+      </Button>
+      <ModalSubmitForm
+        title="Submit Task For Mentor Review"
+        data={state}
+        submitted={state?.submitted}
+        submit={saveSolution}
+        close={closeModal}
+        errorText={state?.errorText}
+        loading={state?.loading}
+      >
+        <CourseTaskSelect groupBy="deadline" data={state?.data?.courseTasks ?? []} />
+        <Form.Item
+          label="Add a solution link"
+          name="url"
+          required
+          rules={[{ message: 'Please enter valid URL', pattern: urlPattern }]}
+        >
+          <Input />
+        </Form.Item>
+      </ModalSubmitForm>
+    </>
+  );
+}
+
+export default SubmitTaskSolution;
