+++ client/src/modules/StudentDashboard/components/SubmitTaskSolution/index.ts
@@ -1,0 +1,1 @@
+export { default as SubmitTaskSolution } from './SubmitTaskSolution';
